feat: basic user functions
